Validate relationships format as part of the FormatRequired plug
The spec does not appear, to me, to be explicit about the behavior
around PATCH and is only very explicit about POST behavior. The examples
in the docs do suggest PATCH may be very similar to POST but I'm not
sure if it's safe to presume that all users of this library draw the
same implications I do from that part of the spec.