Clarify output of ErrorView raw attribute functions
Uses a new `error_attrs` type to represent the raw output of
`build_error`. This is the output of some of the newer error functions.

Also includes an `error_attrs` suffix on those functions to highlight
that they return the raw `build_error` output structure instead of the
JSON:API-friendly error structure returned by most of the other
ErrorView functions.